+++ SWLOR.Game.Server/ValueObject/CraftingData.cs
@@ -7,7 +7,7 @@ namespace SWLOR.Game.Server.ValueObject
 {
     public class CraftingData
     {
-        public int BlueprintID { get; set; }
+        public long BlueprintID { get; set; }
         public CraftBlueprint Blueprint { get; set; }
         public int MainMinimum { get; set; }
         public int MainMaximum { get; set; }
